re-instate server initializer; remove unused classes
