+++ src/main/scala/org/broadinstitute/dsde/agora/server/Agora.scala
@@ -7,7 +7,7 @@ object ProductionAgora extends Agora {
 }
 
 class Agora() extends LazyLogging with App {
-  lazy val server = new ServerInitializer2()
+  lazy val server = new ServerInitializer()
 
   sys addShutdownHook stop()
